MODINVSTOR-492: Refactor tests
